Debuging

+++ src/main/java/com/storm/iotdata/Bolt_sum.java
@@ -109,7 +109,7 @@ class Bolt_sum extends BaseRichBolt {
             Integer house_id     = (Integer) tuple.getValueByField("house_id");
             Double  value        = (Double) tuple.getValueByField("value");
             String household_deviceid = (String)tuple.getValueByField("household_deviceid");
-            int slice_num = (Integer)tuple.getValueByField("slice_num");
+            Long slice_num = (Long)tuple.getValueByField("slice_num");
             String date = (String)tuple.getValueByField("year") + "/" + (String)tuple.getValueByField("month") + "/" + (String)tuple.getValueByField("day");
             String slice_name = date + " " +  String.format("%02d", Math.floorDiv((slice_num*windows),60)) + ":" +  String.format("%02d", (slice_num*windows)%60) + "->" +  String.format("%02d", Math.floorDiv(((slice_num+1)*windows),60)) + ":" +  String.format("%02d", ((slice_num+1)*windows)%60) ;
             HashMap<String, HashMap<String, Double>> data_house = data.getOrDefault(house_id, new HashMap<String, HashMap<String, Double>>());
